Add files via upload
It is very simple logic but helpful. similarly, for any reason, you can write a message or anything. you can type in lowercase but you need of uppercase. using this application it is very simple, just copy your message then open this application and choose your option and paste your message and done.